% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{caliBISG-data}
\alias{caliBISG-data}
\alias{calibisg-data}
\alias{download_data}
\alias{delete_data}
\alias{available_data}
\alias{load_data}
\title{Download caliBISG data files}
\usage{
download_data(state, year, progress = TRUE, overwrite = FALSE)

delete_data(state, year)

available_data(year)

load_data(state, year = 2020)
}
\arguments{
\item{state}{(character vector) For \code{download_data()} or \code{delete_data()}, the
states to download or delete. The default is all available states. If
specifying particular states, they should be provided as two-letter
abbreviations. For \code{load_data()}, a single state to load.}

\item{year}{(integer vector) For \code{download_data()} or \code{delete_data()}, the
years to download or delete. The default is all available years. For
\code{load_data()}, a single year to load. For \code{available_data()}, a single year
to filter the available data files. If not specified, all years are
considered.}

\item{progress}{(logical) Whether to show a progress bar while downloading
the data. The default is \code{TRUE}.}

\item{overwrite}{(logical) For \code{download_data()}, whether to overwrite
existing files if files with the same names already exist. The default is
\code{FALSE}.}
}
\value{
\itemize{
\item \code{download_data()}: (logical) \code{TRUE}, invisibly, if no error.
}

\itemize{
\item \code{delete_data()}: (logical) Invisibly, a vector indicating if each specified
file has been deleted successfully.
}

\itemize{
\item \code{available_data()}: (character vector) The names of the data files
that have already been downloaded.
}

\itemize{
\item \code{load_data()}: (data frame) The caliBISG data for the specified state and year.
}
}
\description{
Currently caliBISG (Greengard and Gelman, 2025) is available for
the following states and years:
\itemize{
\item States: FL, GA, NC, NY, OH, OK, VT, WA
\item Years: 2020
}

Each state-year combination has a separate data file that is used
internally by the package when the user requests caliBISG estimates for
that state and year. Because most of the files are hundreds of megabytes
each, they are not included with the package by default and must instead be
downloaded from GitHub. See \strong{Details}.

We are working on adding additional states and years. When caliBISG is
unavailable we still provide traditional BISG, which does not require
downloading any files.
}
\details{
\itemize{
\item \code{download_data()}: Download the CSV data files from GitHub for the
specified states and years. The downloaded CSV files are converted to data
frames and stored as \link[base:readRDS]{RDS} files internally in a
package-specific data \link[tools:userdir]{directory}. A GitHub personal
access token (PAT) is not required but can be used to increase rate limits.
If the \strong{gitcreds} package is installed and has a credential for
\code{github.com}, the token provided by \code{\link[gitcreds:gitcreds_get]{gitcreds::gitcreds_get()}} will be used.
Otherwise the environment variables \code{GITHUB_PAT} and \code{GITHUB_TOKEN} will be
checked in that order. If no token is found, the request will be made
anonymously.
}

\itemize{
\item \code{delete_data()}: Delete all or a subset of the data files stored internally.
}

\itemize{
\item \code{available_data()}: List the names of the data files that have already been
downloaded and are available for use.
}

\itemize{
\item \code{load_data()}: Load the data for a particular state and  year. This is only
necessary if you want to work with the full data files directly. When using
the functions provided by this package (e.g. \code{\link[=race_probabilities]{race_probabilities()}}) the data
will be loaded internally automatically.
}
}
\examples{
\dontrun{
# Download data just for Oklahoma and Vermont for 2020
download_data(state = c("OK", "VT"), year = 2020)

# Download all available data
download_data()

# List the data files that have already been downloaded
available_data()

# Delete the downloaded data for OK
delete_data("OK")

# Delete all downloaded data
delete_data()
}
}
\references{
Philip Greengard and Andrew Gelman (2025). A calibrated BISG for
inferring race from surname and geolocation.
\emph{Journal of the Royal Statistical Society Series A: Statistics in Society}.
\doi{10.1093/jrsssa/qnaf003}.
}
