% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliBISG-package.R
\docType{package}
\name{caliBISG-package}
\alias{caliBISG}
\alias{caliBISG-package}
\title{caliBISG: A Calibrated BISG for Inferring Race from Surname and Geolocation}
\description{
Bayesian improved surname geocoding (BISG) is a ubiquitous tool for predicting race and ethnicity using an individual’s geolocation and surname, however statistical dependence of surname and geolocation within racial/ethnic categories in the United States results in biases for minority subpopulations. Based on the method proposed by Greengard and Gelman (2025) \doi{10.1093/jrsssa/qnaf003}, we augment the data used by traditional BISG (distributions of race by geolocation and race by surname) with the distribution of surname by geolocation obtained from state voter files. We then use a raking-based algorithm to compute calibrated BISG (caliBISG) predictions. caliBISG is easily interpretable, more accurate, and better calibrated than traditional BISG.
}
\references{
Philip Greengard and Andrew Gelman (2025). A calibrated BISG for
inferring race from surname and geolocation.
\emph{Journal of the Royal Statistical Society Series A: Statistics in Society}.
\doi{10.1093/jrsssa/qnaf003}.

2020 Census State Redistricting Data (Public Law 94-171) Summary File
Prepared by the U.S. Census Bureau, 2021. Retrived from
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}.

U.S. Census Bureau. (2016). Frequently Occurring Surnames from the 2010
Census. Retrieved from
\url{https://www.census.gov/data/developers/data-sets/surnames.html}.
}
\author{
\strong{Maintainer}: Jonah Gabry \email{jsg2201@columbia.edu}

Authors:
\itemize{
  \item Philip Greengard
  \item Gustavo Novoa
  \item Andrew Gelman
}

}
\keyword{internal}
