% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{caliBISG-predict}
\alias{caliBISG-predict}
\alias{calibisg-predict}
\alias{most_probable_race}
\alias{race_probabilities}
\alias{print.compare_calibisg}
\alias{valid_counties}
\alias{fips_to_county}
\title{Predict race given surname and location}
\usage{
most_probable_race(name, state, county, year = 2020)

race_probabilities(name, state, county, year = 2020)

\method{print}{compare_calibisg}(
  x,
  ...,
  digits = getOption("calibisg.digits", 2),
  max_print = getOption("calibisg.max_print", 4)
)

valid_counties(state, year = 2020)

fips_to_county(fips, year = 2020)
}
\arguments{
\item{name}{(character vector) A vector of surnames. Coerced to lowercase
internally.}

\item{state}{(character vector) A vector of state abbreviations. Coerced to
uppercase internally. If a single state is provided, it is recycled to
match the length of \code{name}.}

\item{county}{(character vector) A vector of counties. Coerced to lowercase
internally. If a single county is provided, it is recycled to match the
length of \code{name}. To use FIPS codes instead of county names, use the
\code{fips_to_county()} function to convert FIPS codes to county names first.}

\item{year}{(integer) The year of the data to use to compute the estimates.
The default is \code{2020}, which is currently the only available year. This
default may change in the future when more years become available.}

\item{x}{(compare_calibisg) For \code{print()}, the object returned by
\code{race_probabilities()}, which is a data frame with subclass
\code{"compare_calibisg"}.}

\item{...}{Currently unused.}

\item{digits}{(integer) For \code{print()}, the number of digits to display in the
output. The default is to use two digits unless the global option
\code{calibisg.digits} has been set.}

\item{max_print}{(integer) For \code{print()}, the maximum number of rows to
print. Because the tables take up a lot of space in the console, the
default is to print at most four tables unless the global option
\code{calibisg.max_print} has been set.}

\item{fips}{(character vector) A vector of 5-digit FIPS codes.}
}
\value{
\itemize{
\item \code{most_probable_race()}: (data frame) A data frame with number of rows equal
to the length of the input vectors and the following columns:
\itemize{
\item \code{name} (string): The surname.
\item \code{year} (integer): The year of the data used to compute the estimates.
\item \code{state} (string): The state.
\item \code{county} (string): The county.
\item \code{calibisg_race} (string): The most probable race according to caliBISG.
\item \code{bisg_race} (string): The most probable race according to traditional BISG.
\item \code{in_census} (logical): Whether the surname is found in the list of
names that appear at least 100 times in the census.
}
}

\itemize{
\item \code{race_probabilities()}: (data frame) A data frame with number of rows equal
to the length of the input vectors and the same columns as
\code{most_probable_race()}, except the \code{calibisg_race} and \code{bisg_race} columns
are each replaced by multiple columns giving the probabilities of the various
races, not just the single most probable race. Those columns are:
\itemize{
\item \code{calibisg_aian} (numeric): The caliBISG estimate for American Indian and Alaskan Native.
\item \code{bisg_aian} (numeric): The traditional BISG estimate for American Indian and Alaskan Native.
\item \code{calibisg_api} (numeric): The caliBISG estimate for Asian and Pacific Islander.
\item \code{bisg_api} (numeric): The traditional BISG estimate for Asian and Pacific Islander.
\item \code{calibisg_black_nh} (numeric): The caliBISG estimate for non-Hispanic Black.
\item \code{bisg_black_nh} (numeric): The traditional BISG estimate for non-Hispanic Black.
\item \code{calibisg_hispanic} (numeric): The caliBISG estimate for Hispanic.
\item \code{bisg_hispanic} (numeric): The traditional BISG estimate for Hispanic.
\item \code{calibisg_white_nh} (numeric): The caliBISG estimate for non-Hispanic White.
\item \code{bisg_white_nh} (numeric): The traditional BISG estimate for non-Hispanic
\item \code{calibisg_other} (numeric): The caliBISG estimate for other.
\item \code{bisg_other} (numeric): The traditional BISG estimate for other.
}

The data frame also has class \code{"compare_calibisg"}, which enables defining a
custom \code{print()} method.
}

\itemize{
\item \code{print()}: The input, invisibly.
}

\itemize{
\item \code{valid_counties()}: (character vector) County names.
}

\itemize{
\item \code{fips_to_county()}: (character vector) County names.
}
}
\description{
Predict race given surname and location using the calibrated
BISG (caliBISG) method from Greengard and Gelman (2025). Traditional BISG
estimates are also provided. For some state, county, and surname
combinations the caliBISG estimate will not be available. In those cases we
still provide traditional BISG estimates as long as the state and county
are valid.

Before caliBISG is available, the data files for the relevant states and
years must be downloaded using \code{\link[=download_data]{download_data()}}.

The first query to \code{most_probable_race()} or \code{race_probabilities()} for a
particular state and year may take a few seconds to first load the relevant
caliBISG data internally. Subsequent queries for the same state and year
will be faster.

If the \code{state} argument to \code{most_probable_race()} or \code{race_probabilities()}
contains multiple states, they can be processed in parallel via the
\pkg{future} package by setting a \link[future:plan]{future::plan}.
}
\details{
\itemize{
\item \code{most_probable_race()}: Get the single most probable race given
surname and location.
}

\itemize{
\item \code{race_probabilities()}: Get probabilities for all of the races given
surname and location, rather than the single most probable race.
}

\itemize{
\item \code{print()}: Pretty print the output of \code{race_probabilities()}, making it
easier to compare caliBISG and BISG estimates. It prints a separate
table for each row in the returned data frame up to \code{max_print} rows.
}

\itemize{
\item \code{valid_counties()}: List the valid county names for a given state and year.
}

\itemize{
\item \code{fips_to_county()}: Convert 5-digit FIPS codes to county names.
}
}
\examples{
\dontrun{
download_data(c("VT", "OK"), 2020)

most_probable_race("Smith", "OK", "Tulsa")
most_probable_race(
  name = c("Lopez", "Jackson"),
  state = c("VT", "OK"),
  county = c("Chittenden", "Tulsa")
)

race_probabilities("Smith", "OK", "Tulsa")
race_probabilities("Lopez", "VT", "Chittenden")
probs2 <- race_probabilities(
  name = c("Lopez", "Smith"),
  state = c("VT", "OK"),
  county = c("Chittenden", "Tulsa")
)
str(probs2)
print(probs2, digits = 3)

# caliBISG is not yet available for RI but we can still get
# regular BISG if we input a valid county
valid_counties("RI")
race_probabilities("Jones", "RI", "Providence")

# using FIPS codes
# we can specify Albany County either by name or by its code 36001
most_probable_race(
 name = "Chan",
 state = "NY",
 county = "Albany"
)
most_probable_race(
 name = "Chan",
 state = "NY",
 county = fips_to_county("36001")
)
}


}
\references{
Philip Greengard and Andrew Gelman (2025). A calibrated BISG for
inferring race from surname and geolocation.
\emph{Journal of the Royal Statistical Society Series A: Statistics in Society}.
\doi{10.1093/jrsssa/qnaf003}.
}
