% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{download_data}
\alias{download_data}
\alias{load_data}
\alias{data_dir}
\alias{available_data}
\alias{delete_all_data}
\title{Download data}
\usage{
download_data(states = NULL, ...)

load_data(state)

data_dir()

available_data()

delete_all_data()
}
\arguments{
\item{states}{(character vector) The states to download data for. The default
(\code{NULL}) is to download for all states. States should be provided as
two-letter state \link[datasets:state]{abbreviations}.}

\item{...}{Optional arguments (other than \code{url} and \code{destfile}) passed to
\code{\link[utils:download.file]{utils::download.file()}}.}

\item{state}{(string) For \code{load_data()}, the state to load.}
}
\value{
\itemize{
\item \code{download_data()}: (character vector) The path(s) to the downloaded
data file(s), invisibly.
}

\itemize{
\item \code{load_data()}: (data frame) A data frame of the data for the
specified state.
}

\itemize{
\item \code{data_dir()}: (string) The path to the data directory.
}

\itemize{
\item \code{available_data()}: (character vector) The names of the available
data files.
}
}
\description{
Various functions for downloading and loading data. See
\strong{Details}.
}
\details{
\itemize{
\item \code{download_data()}: Download the required data files for the specified
states. The files will be stored internally in a package-specific data
\link[tools:userdir]{directory}.
\item \code{load_data()}: Load the data for a particular state. This is only necessary
if you want to work with the full data files directly. When using the
functions provided by this package (e.g. \code{\link[=race_probabilities]{race_probabilities()}}) the data
will be loaded internally automatically.
\item \code{data_dir()}: Get the path to the internal data directory.
\item \code{available_data()}: List the names of the available data files.
\item \code{delete_all_data()}: Delete all the data files stored internally.
}
}
