% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{most_probable_race}
\alias{most_probable_race}
\alias{race_probabilities}
\alias{compare_race_probabilities}
\alias{print_comparison_tables}
\title{Predict race given surname and location}
\usage{
most_probable_race(name, state, county, year = 2020)

race_probabilities(name, state, county, year = 2020)

compare_race_probabilities(name, state, county, year = 2020)

print_comparison_tables(x, ..., digits = 4)
}
\arguments{
\item{name}{(character vector) For \code{most_probable_race()} and
\code{race_probabilities()}, a single surname. Can be a vector of surnames for
\code{compare_race_probabilities()}. Coerced to lowercase internally.}

\item{state}{(character vector) For \code{most_probable_race()} and
\code{race_probabilities()}, a single state abbreviation. Can be a vector of state
abbreviations for \code{compare_race_probabilities()}.}

\item{county}{(character vector) For \code{most_probable_race()} and
\code{race_probabilities()}, a single county. Can be a vector of counties for
\code{compare_race_probabilities()}. Coerced to lowercase internally.}

\item{year}{(integer) The year of the data used to compute the estimates.
Currently only 2020 is available.}

\item{...}{Currently unused.}

\item{digits}{(integer) For \code{print_comparison_tables()}, the number of digits
to display in the output.}
}
\value{
\itemize{
\item \code{most_probable_race()}: (data frame) A data frame with the following columns:
\itemize{
\item \code{name} (string): The surname.
\item \code{year} (integer): The year of the data used to compute the estimates.
\item \code{state} (string): The state.
\item \code{county} (string): The county.
\item \code{race} (string): The most probable race based on the raking estimates.
}
\item \code{race_probabilities()}: (data frame) A data frame with one
row and the following columns:
\itemize{
\item \code{name} (string): The surname.
\item \code{year} (integer): The year of the data used to compute the estimates.
\item \code{state} (string): The state.
\item \code{county} (string): The county.
\item \code{rake_nh_aian} (numeric): The raking estimate for non-Hispanic American Indian and Alaskan Native.
\item \code{rake_nh_api} (numeric): The raking estimate for non-Hispanic Asian and Pacific Islander.
\item \code{rake_nh_black} (numeric): The raking estimate for non-Hispanic Black.
\item \code{rake_hispanic} (numeric): The raking estimate for Hispanic.
\item \code{rake_nh_white} (numeric): The raking estimate for non-Hispanic White.
\item \code{rake_other} (numeric): The raking estimate for other.
}
\item \code{compare_race_probabilities}: (data frame) Same as \code{race_probabilities()}
but with the following additional columns:
\itemize{
\item \code{bisg_nh_aian} (numeric): The improved BISG estimate for non-Hispanic American Indian and Alaskan Native.
\item \code{bisg_nh_api} (numeric): The improved BISG estimate for non-Hispanic Asian and Pacific Islander.
\item \code{bisg_nh_black} (numeric): The improved BISG estimate for non-Hispanic Black.
\item \code{bisg_hispanic} (numeric): The improved BISG estimate for Hispanic.
\item \code{bisg_nh_white} (numeric): The improved BISG estimate for non-Hispanic White.
\item \code{bisg_other} (numeric): The improved BISG estimate for other.
}
}
}
\description{
Predict race given surname and location using the improved
raking-based BISG method from Greengard and Gelman (2024). See \strong{Details}.

Before these functions can be used, the data files for the relevant states
must be downloaded using \code{\link[=download_data]{download_data()}}.

These functions are just convenience functions for querying the large data
files. You can also \link[=load_data]{load} and use the data files directly.
}
\details{
\itemize{
\item \code{most_probable_race()}: The single most probable race according to the
raking-based approach.
\item \code{race_probabilities()}: Raking-based probabilities for all of the races.
\item \code{compare_race_probabilities()}: Same as \code{race_probabilities()} but also
includes "Improved BISG" estimates. Here, "Improved BISG" refers to
traditional BISG but including an adjustment to those predictions that
accounts for the fact that we’re making predictions on registered voters, not
the general population. It uses the registered voter status x race
distribution and an application of Bayes rule.
\item \code{print_comparison_tables()}: Pretty print the output of
\code{compare_race_probabilities()}. Prints a separate table for each input
record, so this is most useful when only a small number of records were
requested.
}
}
\examples{
\dontrun{
most_probable_race("smith", "wa", "king")
most_probable_race("lopez", "nc", "burke")

race_probabilities("smith", "wa", "king")
race_probabilities("lopez", "nc", "burke")

comp1 <- compare_race_probabilities("smith", "wa", "king")
str(comp1)
print_comparison_tables(comp1)

comp2 <- compare_race_probabilities(
  name = c("Lopez", "Jackson"),
  state = c("NC", "WA"),
  county = c("Burke", "King")
)
str(comp2)
print_comparison_tables(comp2)
}

}
\references{
Philip Greengard and Andrew Gelman (2024). An improved BISG for
inferring race from surname and geolocation.
\url{https://arxiv.org/abs/2304.09126}.
}
