% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_bisg.R
\name{predict_bisg}
\alias{predict_bisg}
\title{BISG prediction from vectors of names + counties}
\usage{
predict_bisg(name, county, state, year = 2020)
}
\arguments{
\item{name}{(character vector) A vector of surnames. Coerced to lowercase
internally.}

\item{county}{(character vector) A vector of counties. Coerced to lowercase
internally.}

\item{state}{(string) A state abbreviation.}

\item{year}{(integer) The year of the data used to compute the estimates.
Currently only 2020 is available.}
}
\value{
(data frame) A data frame with colums \code{name}, \code{year}, \code{state},
\code{county}, plus six BISG probability columns:
* \code{bisg_aian}
* \code{bisg_api}
* \code{bisg_black_nh}
* \code{bisg_hispanic}
* \code{bisg_white_nh}
* \code{bisg_other}
}
\description{
Compute Bayesian Improved Surname Geocoding (BISG) race‐probability
estimates using census surname and county reference tables.
}
\keyword{internal}
