% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{race_probabilities}
\alias{race_probabilities}
\alias{compare_race_probabilities}
\alias{predict_race}
\alias{print.raking_bisg}
\title{Predict race given surname and location}
\usage{
race_probabilities(name, state, county, year = 2020)

compare_race_probabilities(name, state, county, year = 2020)

predict_race(name, state, county, year = 2020)

\method{print}{raking_bisg}(x, ..., digits = 4)
}
\arguments{
\item{name}{(string) A surname. Will be coerced to lowercase internally.}

\item{state}{(string) A two-letter state abbreviation.}

\item{county}{(string) A county within \code{state}. Will be coerced to lowercase
internally.}

\item{year}{(integer) The year of the data used to compute the estimates.
Currently only 2020 is available.}

\item{...}{Currently unused.}

\item{digits}{(integer) For printing the output of
\code{compare_race_probabilities()}, the number of digits to display in the
output.}
}
\value{
\itemize{
\item \code{race_probabilities()}: (data frame) A data frame with one
row and the following columns:
\itemize{
\item \code{name} (string): The surname.
\item \code{year} (integer): The year of the data used to compute the estimates.
\item \code{state} (string): The state.
\item \code{county} (string): The county.
\item \code{rake_nh_aian} (numeric): The raking estimate for non-Hispanic American Indian and Alaskan Native.
\item \code{rake_nh_api} (numeric): The raking estimate for non-Hispanic Asian and Pacific Islander.
\item \code{rake_nh_black} (numeric): The raking estimate for non-Hispanic Black.
\item \code{rake_hispanic} (numeric): The raking estimate for Hispanic.
\item \code{rake_nh_white} (numeric): The raking estimate for non-Hispanic White.
\item \code{rake_other} (numeric): The raking estimate for other.
}
\item \code{compare_race_probabilities}: (data frame) Same as \code{race_probabilities()}
but the data frame has a custom class and print method and includes the
following additional columns:
\itemize{
\item \code{bisg_nh_aian} (numeric): The improved BISG estimate for non-Hispanic American Indian and Alaskan Native.
\item \code{bisg_nh_api} (numeric): The improved BISG estimate for non-Hispanic Asian and Pacific Islander.
\item \code{bisg_nh_black} (numeric): The improved BISG estimate for non-Hispanic Black.
\item \code{bisg_hispanic} (numeric): The improved BISG estimate for Hispanic.
\item \code{bisg_nh_white} (numeric): The improved BISG estimate for non-Hispanic White.
\item \code{bisg_other} (numeric): The improved BISG estimate for other.
}
\item \code{predict_race()}: (data frame) A data frame with the following columns:
\itemize{
\item \code{name} (string): The surname.
\item \code{year} (integer): The year of the data used to compute the estimates.
\item \code{state} (string): The state.
\item \code{county} (string): The county.
\item \code{race} (string): The most probable race based on the raking estimates.
}
}
}
\description{
Predict race given surname and location using the improved
raking-based BISG method from Greengard and Gelman (2024). See \strong{Details}.

Before these functions can be used, the data files for the relevant states
must be downloaded using \code{\link[=download_data]{download_data()}}.

These functions are just convenience functions for querying the large data
files. You can also \link[=load_data]{load} and use the data files directly.
}
\details{
\itemize{
\item \code{race_probabilities()}: Loads the data internally for the specified
state-year, then looks up the surname and county and returns the
raking-based probabilities for all of the races.
\item \code{compare_race_probabilities()}: Same as \code{race_probabilities()} but also
includes "Improved BISG" estimates and has a custom print method that
produces a table comparing the two sets of estimates. Here, "Improved BISG"
refers to traditional BISG but including an adjustment to those predictions
that accounts for the fact that we’re making predictions on registered
voters, not the general population. It uses the registered voter status x
race distribution and an application of Bayes rule.
\item \code{predict_race()}: Same as \code{race_probabilities()} but only includes the
single most probable race, not probabilities for all the races.
}
}
\examples{
\dontrun{
predict_race("smith", "wa", "king")
predict_race("lopez", "nc", "burke")

race_probabilities("smith", "wa", "king")
race_probabilities("lopez", "nc", "burke")

compare_race_probabilities("smith", "wa", "king")
compare_race_probabilities("lopez", "nc", "burke")

# Only use 2 digits when printing the comparison table
print(
 compare_race_probabilities("lopez", "nc", "burke"),
 digits = 2
)
}

}
\references{
Philip Greengard and Andrew Gelman (2024). An improved BISG for
inferring race from surname and geolocation.
\url{https://arxiv.org/abs/2304.09126}.
}
