% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{set_temporary_local_directory}
\alias{set_temporary_local_directory}
\title{Set and the temporarly local path to the data files. Will replace this
with downloading them eventually.}
\usage{
set_temporary_local_directory(dir = NULL)
}
\arguments{
\item{dir}{Path to local directory.}
}
\description{
Set and the temporarly local path to the data files. Will replace this
with downloading them eventually.
}
\keyword{internal}
