% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisg.R
\name{bisg}
\alias{bisg}
\title{Compute traditional BISG}
\usage{
bisg(name, state, county, year = 2020)
}
\arguments{
\item{name}{(character vector) A vector of surnames. Coerced to lowercase
internally.}

\item{state}{(character vector) A vector of state abbreviations. Coerced to
uppercase internally. If a single state is provided, it is recycled to
match the length of \code{name}.}

\item{county}{(character vector) A vector of counties. Coerced to lowercase
internally. If a single county is provided, it is recycled to match the
length of \code{name}. To use FIPS codes instead of county names, use the
\code{fips_to_county()} function to convert FIPS codes to county names first.}

\item{year}{(integer) The year of the data to use to compute the estimates.
The default is \code{2020}, which is currently the only available year. This
default may change in the future when more years become available.}
}
\value{
(data frame) A data frame with colums \code{name}, \code{year}, \code{state},
\code{county}, plus six BISG probability columns:
\itemize{
\item \code{bisg_aian}
\item \code{bisg_api}
\item \code{bisg_black_nh}
\item \code{bisg_hispanic}
\item \code{bisg_white_nh}
\item \code{bisg_other}
}
}
\description{
Compute Bayesian Improved Surname Geocoding (BISG) estimates
using census surname and county reference tables.
}
\examples{
\dontrun{
bisg(
  name = c("Lopez", "Jackson", "Smith"),
  county = c("King", "King", "Chittenden"),
  state = c("WA", "WA", "VT"),
  year = 2020
)
}

}
\references{
2020 Census State Redistricting Data (Public Law 94-171) Summary File
Prepared by the U.S. Census Bureau, 2021. Retrived from
\url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}.

U.S. Census Bureau. (2016). Frequently Occurring Surnames from the 2010
Census. Retrieved from
\url{https://www.census.gov/data/developers/data-sets/surnames.html}.
}
\keyword{internal}
