% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{caliBISG-data}
\alias{caliBISG-data}
\alias{calibisg-data}
\alias{download_data}
\alias{load_data}
\alias{available_data}
\alias{delete_all_data}
\title{Download caliBISG data files}
\usage{
download_data(state, year, progress = TRUE)

load_data(state, year = 2020)

available_data()

delete_all_data()
}
\arguments{
\item{state}{(character vector) For \code{download_data()}, the states to
download. The default is to download caliBISG data for all available
states. If specifying particular states, they should be provided as
two-letter abbreviations. For \code{load_data()}, a single state to load.}

\item{year}{(integer vector) For \code{download_data()}, the years to download.
The default is to download caliBISG data for all available years. For
\code{load_data()}, a single year to load.}

\item{progress}{(logical) Whether to show a progress bar while downloading
the data. The default is \code{TRUE}.}
}
\value{
\itemize{
\item \code{download_data()}: (logical) \code{TRUE}, invisibly, if no error.
}

\itemize{
\item \code{load_data()}: (data frame) The caliBISG data for the specified state and year.
}

\itemize{
\item \code{available_data()}: (character vector) The names of the data files
that have already been downloaded.
}

\itemize{
\item \code{delete_all_data()}: (logical) \code{TRUE} or \code{FALSE}, invisibly,
indicating success or failure.
}
}
\description{
Various functions for downloading and loading the caliBISG data
files used by the package. Currently caliBISG is available for the
following states and years:
\itemize{
\item States: FL, GA, NC, NY, OH, OK, VT, WA
\item Years: 2020
}

We are working on adding additional states and years. When caliBISG is
unavailable we still provide traditional BISG, which does not require
downloading any files.
}
\details{
\itemize{
\item \code{download_data()}: Download the CSV data files from GitHub for the
specified states and years. The downloaded files are converted to data frames
and stored as \link[base:readRDS]{RDS} files internally in a package-specific
data \link[tools:userdir]{directory}. A GitHub personal access token (PAT) can
be used to increase rate limits. If the \strong{gitcreds} package is installed and
has a credential for \code{github.com}, the token provided by
\code{\link[gitcreds:gitcreds_get]{gitcreds::gitcreds_get()}} will be used. Otherwise the environment variables
\code{GITHUB_PAT} and \code{GITHUB_TOKEN} will be checked in that order. If no token is
found, the request will be made anonymously, in which case you may run into
GitHub's unauthenticated rate limit if you are trying to download many files
within one hour.
}

\itemize{
\item \code{load_data()}: Load the data for a particular state and  year. This is only
necessary if you want to work with the full data files directly. When using
the functions provided by this package (e.g. \code{\link[=race_probabilities]{race_probabilities()}}) the data
will be loaded internally automatically.
}

\itemize{
\item \code{available_data()}: List the names of the data files that have already been
downloaded and are available for use.
}

\itemize{
\item \code{delete_all_data()}: Delete all the data files stored internally.
}
}
\examples{
\dontrun{
# Download data just for Oklahoma and Washington for 2020
download_data(state = c("OK", "WA"), year = 2020)

# Download all available data
download_data()

# List the data files that have already been downloaded
available_data()
}
}
