% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{most_probable_race}
\alias{most_probable_race}
\alias{race_probabilities}
\alias{print_comparison_tables}
\title{Predict race given surname and location}
\usage{
most_probable_race(name, state, county, year = 2020)

race_probabilities(name, state, county, year = 2020)

print_comparison_tables(x, ..., digits = 2)
}
\arguments{
\item{name}{(character vector) A vector of surnames. Coerced to lowercase
internally.}

\item{state}{(character vector) A vector of state abbreviations. Coerced to
uppercase internally.}

\item{county}{(character vector) A vector of counties. Coerced to lowercase
internally.}

\item{year}{(integer) The year of the data to use to compute the estimates.
The default is 2020, which is currently the only available year. This
default may change in the future when more years become available.}

\item{x}{For \code{print_comparison_tables()}, the object returned by
\code{race_probabilities()}.}

\item{...}{Currently unused.}

\item{digits}{(integer) For \code{print_comparison_tables()}, the number of digits
to display in the output. The default is \code{2}.}
}
\value{
\itemize{
\item \code{most_probable_race()}: (data frame) A data frame with the following columns:
\itemize{
\item \code{name} (string): The surname.
\item \code{year} (integer): The year of the data used to compute the estimates.
\item \code{state} (string): The state.
\item \code{county} (string): The county.
\item \code{calibisg_race} (string): The most probable race according to caliBISG.
\item \code{bisg_race} (string): The most probable race according to traditional BISG.
\item \code{in_census} (logical): Whether the surname is found in the list of
names that appear at least 100 times in the census.
}
\item \code{race_probabilities()}: (data frame) A data frame with the same columns
as \code{most_probable_race()}, except the \code{calibisg_race} and \code{bisg_race} columns
are each replaced by multiple columns giving the probabilities of the various
races, not just the single most probable race. The columns are:
\itemize{
\item \code{calibisg_aian} (numeric): The caliBISG estimate for American Indian and Alaskan Native.
\item \code{bisg_aian} (numeric): The traditional BISG estimate for American Indian and Alaskan Native.
\item \code{calibisg_api} (numeric): The caliBISG estimate for Asian and Pacific Islander.
\item \code{bisg_api} (numeric): The traditional BISG estimate for Asian and Pacific Islander.
\item \code{calibisg_black_nh} (numeric): The caliBISG estimate for non-Hispanic Black.
\item \code{bisg_black_nh} (numeric): The traditional BISG estimate for non-Hispanic Black.
\item \code{calibisg_hispanic} (numeric): The caliBISG estimate for Hispanic.
\item \code{bisg_hispanic} (numeric): The traditional BISG estimate for Hispanic.
\item \code{calibisg_white_nh} (numeric): The caliBISG estimate for non-Hispanic White.
\item \code{bisg_white_nh} (numeric): The traditional BISG estimate for non-Hispanic
\item \code{calibisg_other} (numeric): The caliBISG estimate for other.
\item \code{bisg_other} (numeric): The traditional BISG estimate for other.
}
}
}
\description{
Predict race given surname and location using the calibrated
BISG (caliBISG) method from Greengard and Gelman (2025). Traditional BISG
estimates are also provided. See \strong{Details}.

Before caliBISG is available, the data files for the relevant states and
years must be downloaded using \code{\link[=download_data]{download_data()}}.
}
\details{
The \code{most_probable_race()} function finds the single most probable race given
surname and location. The \code{race_probabilities()} function provides
probabilities for all of the races rather than the single most probable race.
The \code{print_comparison_tables()} pretty prints the output of
\code{race_probabilities()}. It prints a separate table for each input record and
so is most useful when only a small number of records were requested.

The first query for a particular \code{state} and \code{year} may take a few seconds
to first load the caliBISG data internally. Subsequent calls for the same
\code{state} and \code{year} will be faster.

For some state, county, and surname combinations the caliBISG estimate
will not be available. In those cases we still provide traditional BISG
estimates as long as the state and county are valid.
}
\examples{
\dontrun{
download_data(c("VT", "WA"), 2020)

most_probable_race("smith", "wa", "king")
most_probable_race(
  name = c("Lopez", "Jackson"),
  state = c("VT", "WA"),
  county = c("Chittenden", "King")
)

race_probabilities("smith", "wa", "king")
race_probabilities("lopez", "vt", "chittenden")
probs2 <- race_probabilities(
  name = c("Lopez", "Smith"),
  state = c("VT", "WA"),
  county = c("Chittenden", "King")
)
print_comparison_tables(race_probabilities("smith", "wa", "king"))
print_comparison_tables(probs2, digits = 2)
}

}
\references{
Philip Greengard and Andrew Gelman (2025). A calibrated BISG for
inferring race from surname and geolocation.
\emph{Journal of the Royal Statistical Society Series A: Statistics in Society}.
\url{https://doi.org/10.1093/jrsssa/qnaf003}.
}
