% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{most_probable_race}
\alias{most_probable_race}
\alias{race_probabilities}
\alias{compare_race_probabilities}
\alias{print_comparison_tables}
\title{Predict race given surname and location}
\usage{
most_probable_race(name, state, county, year = 2020)

race_probabilities(name, state, county, year = 2020)

compare_race_probabilities(name, state, county, year = 2020)

print_comparison_tables(x, ..., digits = 4)
}
\arguments{
\item{name}{(character vector) A vector of surnames. Coerced to lowercase
internally.}

\item{state}{(character vector) A vector of state abbreviations. Coerced to
uppercase internally.}

\item{county}{(character vector) A vector of counties. Coerced to lowercase
internally.}

\item{year}{(integer) The year of the data used to compute the estimates.
Currently only 2020 is available.}

\item{x}{For \code{print_comparison_tables()}, the object returned by
\code{compare_race_probabilities()}.}

\item{...}{Currently unused.}

\item{digits}{(integer) For \code{print_comparison_tables()}, the number of digits
to display in the output.}
}
\value{
\itemize{
\item \code{most_probable_race()}: (data frame) A data frame with the following columns:
\itemize{
\item \code{name} (string): The surname.
\item \code{year} (integer): The year of the data used to compute the estimates.
\item \code{state} (string): The state.
\item \code{county} (string): The county.
\item \code{race} (string): The most probable race based on the raking estimates.
\item \code{in_census} (logical): Whether the surname is found in the list of
names that appear at least 100 times in the census.
}
\item \code{race_probabilities()}: (data frame) A data frame with the same columns
as \code{most_probable_race()}, except the \code{race} column is replaced with the
following columns that give the raking-based probabilities:
\itemize{
\item \code{rake_aian} (numeric): The raking estimate for American Indian and Alaskan Native.
\item \code{rake_api} (numeric): The raking estimate for Asian and Pacific Islander.
\item \code{rake_black_nh} (numeric): The raking estimate for non-Hispanic Black.
\item \code{rake_hispanic} (numeric): The raking estimate for Hispanic.
\item \code{rake_white_nh} (numeric): The raking estimate for non-Hispanic White.
\item \code{rake_other} (numeric): The raking estimate for other.
}
\item \code{compare_race_probabilities}: (data frame) Same as \code{race_probabilities()}
but with two additional columns for each race, one for the traditional BISG
(prefixed with \code{bisg_}) and one for "improved BISG" (prefixed with
\code{voter_bisg_}). See \strong{Details}.
}
}
\description{
Predict race given surname and location using the raking-based
BISG method from Greengard and Gelman (2024). See \strong{Details}.

Before these functions can be used, the data files for the relevant states
must be downloaded using \code{\link[=download_data]{download_data()}}.

These functions are just convenience functions for querying the large data
files. You can also \link[=load_data]{load} and use the data files directly.

NOTE: the first query for a particular \code{state} and \code{year} may take a few
seconds to first load the data internally. Subsequent calls for the same
\code{state} and \code{year} will be faster.
}
\details{
\itemize{
\item \code{most_probable_race()}: The single most probable race according to the
raking-based approach to BISG.
\item \code{race_probabilities()}: Raking-based probabilities for all of the races.
\item \code{compare_race_probabilities()}: Same as \code{race_probabilities()} but also
includes traditional BISG estimates and "improved" BISG estimates for
comparison purposes. Here, "improved BISG" refers to traditional BISG but
including an adjustment to those predictions that accounts for the fact that
we're making predictions on registered voters, not the general population. It
uses the registered voter status x race distribution and an application of
Bayes rule. Greengard and Gelman (2024) considers the raking-based approach
to be an improvement over both traditional and "improved" BISG.
\item \code{print_comparison_tables()}: Pretty print the output of
\code{compare_race_probabilities()}. Prints a separate table for each input
record, so this is most useful when only a small number of records were
requested.
}
}
\examples{
\dontrun{
most_probable_race("smith", "wa", "king")
most_probable_race(
  name = c("Lopez", "Jackson"),
  state = c("NC", "WA"),
  county = c("Burke", "King")
)

race_probabilities("smith", "wa", "king")
race_probabilities("lopez", "nc", "burke")
probs2 <- race_probabilities(
  name = c("Smith", "Lopez"),
  state = c("NC", "WA"),
  county = c("Burke", "King")
)
str(probs2)

comp1 <- compare_race_probabilities("smith", "wa", "king")
str(comp1)
print_comparison_tables(comp1)

comp2 <- compare_race_probabilities(
  name = c("Lopez", "Jackson"),
  state = c("NC", "WA"),
  county = c("Burke", "King")
)
str(comp2)
print_comparison_tables(comp2)
print_comparison_tables(comp2, digits = 2)
}

}
\references{
Philip Greengard and Andrew Gelman (2024). An improved BISG for
inferring race from surname and geolocation.
\url{https://arxiv.org/abs/2304.09126}.
}
