% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{most_probable_race}
\alias{most_probable_race}
\alias{race_probabilities}
\alias{print.compare_bisg}
\alias{valid_counties}
\title{Predict race given surname and location}
\usage{
most_probable_race(name, state, county, year = 2020)

race_probabilities(name, state, county, year = 2020)

\method{print}{compare_bisg}(
  x,
  ...,
  digits = getOption("calibisg.digits", 2),
  max_print = getOption("calibisg.max_print", 4)
)

valid_counties(state, year = 2020)
}
\arguments{
\item{name}{(character vector) A vector of surnames. Coerced to lowercase
internally.}

\item{state}{(character vector) A vector of state abbreviations. Coerced to
uppercase internally.}

\item{county}{(character vector) A vector of counties. Coerced to lowercase
internally.}

\item{year}{(integer) The year of the data to use to compute the estimates.
The default is \code{2020}, which is currently the only available year. This
default may change in the future when more years become available.}

\item{x}{(compare_bisg) For \code{print()}, the object returned by
\code{race_probabilities()}, which is a data frame with subclass
\code{"compare_bisg"}.}

\item{...}{Currently unused.}

\item{digits}{(integer) For \code{print()}, the number of digits to display in the
output. The default is to use two digits unless the global option
\code{calibisg.digits} has been set.}

\item{max_print}{(integer) For \code{print()}, the maximum number of rows to
print. Because the tables take up a lot of space in the console, the
default is to print at most four tables unless the global option
\code{calibisg.max_print} has been set.}
}
\value{
\itemize{
\item \code{most_probable_race()}: (data frame) A data frame with number of rows equal
to the length of the input vectors and the following columns:
\itemize{
\item \code{name} (string): The surname.
\item \code{year} (integer): The year of the data used to compute the estimates.
\item \code{state} (string): The state.
\item \code{county} (string): The county.
\item \code{calibisg_race} (string): The most probable race according to caliBISG.
\item \code{bisg_race} (string): The most probable race according to traditional BISG.
\item \code{in_census} (logical): Whether the surname is found in the list of
names that appear at least 100 times in the census.
}
\item \code{race_probabilities()}: (data frame) A data frame with number of rows equal
to the length of the input vectors and the same columns as
\code{most_probable_race()}, except the \code{calibisg_race} and \code{bisg_race} columns
are each replaced by multiple columns giving the probabilities of the various
races, not just the single most probable race. Those columns are:
\itemize{
\item \code{calibisg_aian} (numeric): The caliBISG estimate for American Indian and Alaskan Native.
\item \code{bisg_aian} (numeric): The traditional BISG estimate for American Indian and Alaskan Native.
\item \code{calibisg_api} (numeric): The caliBISG estimate for Asian and Pacific Islander.
\item \code{bisg_api} (numeric): The traditional BISG estimate for Asian and Pacific Islander.
\item \code{calibisg_black_nh} (numeric): The caliBISG estimate for non-Hispanic Black.
\item \code{bisg_black_nh} (numeric): The traditional BISG estimate for non-Hispanic Black.
\item \code{calibisg_hispanic} (numeric): The caliBISG estimate for Hispanic.
\item \code{bisg_hispanic} (numeric): The traditional BISG estimate for Hispanic.
\item \code{calibisg_white_nh} (numeric): The caliBISG estimate for non-Hispanic White.
\item \code{bisg_white_nh} (numeric): The traditional BISG estimate for non-Hispanic
\item \code{calibisg_other} (numeric): The caliBISG estimate for other.
\item \code{bisg_other} (numeric): The traditional BISG estimate for other.
}

The data frame also has class \code{"compare_bisg"}, which enables defining a
custom \code{print()} method.
\item \code{valid_counties()}: (character vector) Valid county names for the specified
\code{state} and \code{year}.
}
}
\description{
Predict race given surname and location using the calibrated
BISG (caliBISG) method from Greengard and Gelman (2025). Traditional BISG
estimates are also provided. See \strong{Details}.

Before caliBISG is available, the data files for the relevant states and
years must be downloaded using \code{\link[=download_data]{download_data()}}.
}
\details{
The \code{most_probable_race()} function finds the single most probable race given
surname and location.

The \code{race_probabilities()} function provides probabilities for all of the
races rather than the single most probable race.

The \code{print()} method pretty prints the output of \code{race_probabilities()},
making it easier to compare caliBISG and BISG estimates. It prints a separate
table for each row in the returned data frame up to \code{max_print} rows.

The first query for a particular \code{state} and \code{year} may take a few seconds
to first load the caliBISG data internally. Subsequent calls for the same
\code{state} and \code{year} will be faster.

For some state, county, and surname combinations the caliBISG estimate
will not be available. In those cases we still provide traditional BISG
estimates as long as the state and county are valid. The \code{valid_counties()}
function can be used to check which county names can be specified for a
given state and year.
}
\examples{
\dontrun{
download_data(c("VT", "WA"), 2020)

most_probable_race("Smith", "WA", "King")
most_probable_race(
  name = c("Lopez", "Jackson"),
  state = c("VT", "WA"),
  county = c("Chittenden", "King")
)

race_probabilities("Smith", "WA", "King")
race_probabilities("Lopez", "VT", "Chittenden")
probs2 <- race_probabilities(
  name = c("Lopez", "Smith"),
  state = c("VT", "WA"),
  county = c("Chittenden", "King")
)
str(probs2)
print(probs2, digits = 3)

# caliBISG is not yet available for RI but we can still get
# regular BISG if we input a valid county
valid_counties("RI")
race_probabilities("Jones", "RI", "Providence")
}

}
\references{
Philip Greengard and Andrew Gelman (2025). A calibrated BISG for
inferring race from surname and geolocation.
\emph{Journal of the Royal Statistical Society Series A: Statistics in Society}.
\url{https://doi.org/10.1093/jrsssa/qnaf003}.
}
