% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{most_probable_race}
\alias{most_probable_race}
\alias{race_probabilities}
\alias{print_comparison_tables}
\title{Predict race given surname and location}
\usage{
most_probable_race(name, state, county, year = 2020)

race_probabilities(name, state, county, year = 2020)

print_comparison_tables(x, ..., digits = 4)
}
\arguments{
\item{name}{(character vector) A vector of surnames. Coerced to lowercase
internally.}

\item{state}{(character vector) A vector of state abbreviations. Coerced to
uppercase internally.}

\item{county}{(character vector) A vector of counties. Coerced to lowercase
internally.}

\item{year}{(integer) The year of the data used to compute the estimates.
Currently only 2020 is available.}

\item{x}{For \code{print_comparison_tables()}, the object returned by
\code{compare_race_probabilities()}.}

\item{...}{Currently unused.}

\item{digits}{(integer) For \code{print_comparison_tables()}, the number of digits
to display in the output.}
}
\value{
\itemize{
\item \code{most_probable_race()}: (data frame) A data frame with the following columns:
\itemize{
\item \code{name} (string): The surname.
\item \code{year} (integer): The year of the data used to compute the estimates.
\item \code{state} (string): The state.
\item \code{county} (string): The county.
\item \code{calibisg_race} (string): The most probable race based on the caliBISG estimates.
\item \code{bisg_race} (string): The most probable race based on the caliBISG estimates.
\item \code{in_census} (logical): Whether the surname is found in the list of
names that appear at least 100 times in the census.
}
\item \code{race_probabilities()}: (data frame) A data frame with the same columns
as \code{most_probable_race()}, except the \code{race} column is replaced with the
following columns that give the raking-based probabilities:
\itemize{
\item \code{calibisg_aian} (numeric): The caliBISG estimate for American Indian and Alaskan Native.
\item \code{calibisg_api} (numeric): The caliBISG estimate for Asian and Pacific Islander.
\item \code{calibisg_black_nh} (numeric): The caliBISG estimate for non-Hispanic Black.
\item \code{calibisg_hispanic} (numeric): The caliBISG estimate for Hispanic.
\item \code{calibisg_white_nh} (numeric): The caliBISG estimate for non-Hispanic White.
\item \code{calibisg_other} (numeric): The caliBISG estimate for other.
\item There is also a \verb{bisg_*} column for each race giving the traditional BISG estimate.
}
}
}
\description{
Predict race given surname and location using the raking-based
Calibrated BISG (caliBISG) method from Greengard and Gelman (2024). See
\strong{Details}.

Before these functions can be used, the data files for the relevant states
must be downloaded using \code{\link[=download_data]{download_data()}}.

These functions are just convenience functions for querying the large data
files. You can also \link[=load_data]{load} and use the data files directly.

NOTE: the first query for a particular \code{state} and \code{year} may take a few
seconds to first load the data internally. Subsequent calls for the same
\code{state} and \code{year} will be faster.
}
\details{
\itemize{
\item \code{most_probable_race()}: The single most probable race according to the
caliBISG method and traditional BISG.
\item \code{race_probabilities()}: Probabilities for all of the races based on the
caliBISG method and traditional BISG.
\item \code{print_comparison_tables()}: Pretty print the output of
\code{race_probabilities()}. Prints a separate table for each input
record, so this is most useful when only a small number of records were
requested.
}
}
\examples{
\dontrun{
most_probable_race("smith", "wa", "king")
most_probable_race(
  name = c("Lopez", "Jackson"),
  state = c("VT", "WA"),
  county = c("Chittenden", "King")
)

race_probabilities("smith", "wa", "king")
race_probabilities("lopez", "vt", "chittenden")
probs2 <- race_probabilities(
  name = c("Lopez", "Smith"),
  state = c("VT", "WA"),
  county = c("Chittenden", "King")
)
print_comparison_tables(race_probabilities("smith", "wa", "king"))
print_comparison_tables(probs2, digits = 2)
}

}
\references{
Philip Greengard and Andrew Gelman (2024). An improved BISG for
inferring race from surname and geolocation.
\url{https://arxiv.org/abs/2304.09126}.
}
